# documentation/swagger.json
{"info":{"description":"All api requests for Stuadio hybrid application.","title":"Staudio API","version":"1.0.0"},"openapi":"3.0.0","paths":{"/deal":{"post":{"description":"","parameters":[{"description":"","in":"header","name":"Content-Type","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"successful operation"}},"summary":"Add new deal","tags":["Deals"]}},"/deal/61641fd791f1be141f3dfd62":{"get":{"responses":{"200":{"description":"successful operation"}},"summary":"Get specific deal","tags":["Deals"]},"post":{"description":"","parameters":[{"description":"","in":"header","name":"Content-Type","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"successful operation"}},"summary":"Add new product","tags":["Deals"]}},"/order":{"get":{"responses":{"200":{"description":"successful operation"}},"summary":"Get all orders","tags":["Order"]},"post":{"description":"","parameters":[{"description":"","in":"header","name":"Content-Type","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"successful operation"}},"summary":"Create a new order","tags":["Order"]}},"/order/6171d3126d04e05adbf26cce":{"get":{"responses":{"200":{"description":"successful operation"}},"summary":"Get specific order","tags":["Order"]},"patch":{"description":"","parameters":[{"description":"","in":"header","name":"Content-Type","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"successful operation"}},"summary":"Update order to paid","tags":["Order"]}},"/order/item/6171d3126d04e05adbf26cce":{"post":{"description":"","parameters":[{"description":"","in":"header","name":"Content-Type","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"successful operation"}},"summary":"Create order items","tags":["Order"]}},"/order/user/617142e069b79b137d2ccf33":{"get":{"responses":{"200":{"description":"successful operation"}},"summary":"Get orders from specific user","tags":["Order"]}},"/product":{"get":{"responses":{"200":{"description":"successful operation"}},"summary":"Get all products","tags":["Products"]},"post":{"description":"","parameters":[{"description":"","in":"header","name":"Content-Type","required":true,"schema":{"type":"string"}},{"description":"","in":"header","name":"","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"successful operation"}},"summary":"Add new product","tags":["Products"]}},"/product/6156ed92d170e00abff3dd98":{"get":{"responses":{"200":{"description":"successful operation"}},"summary":"Get specific product","tags":["Products"]}},"/product/review/6156ed92d170e00abff3dd98":{"post":{"description":"","parameters":[{"description":"","in":"header","name":"Content-Type","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"successful operation"}},"summary":"Create new review","tags":["Review"]}},"/user":{"get":{"responses":{"200":{"description":"successful operation"}},"summary":"Get all users","tags":["User"]},"post":{"description":"","parameters":[{"description":"","in":"header","name":"Content-Type","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"successful operation"}},"summary":"Create new user","tags":["User"]}},"/user/617142e069b79b137d2ccf33":{"patch":{"description":"","parameters":[{"description":"","in":"header","name":"Content-Type","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"successful operation"}},"summary":"Update existing user","tags":["User"]}},"/user/authorize":{"post":{"description":"","parameters":[{"description":"","in":"header","name":"Content-Type","required":true,"schema":{"type":"string"}},{"description":"","in":"header","name":"","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"successful operation"}},"summary":"Authorize user","tags":["User"]}}},"servers":[{"url":"http://127.0.0.1:8000/api"}],"tags":[{"description":"","name":"Order"},{"description":"","name":"Deals"},{"description":"","name":"Review"},{"description":"","name":"Products"},{"description":"","name":"User"}]}